requirements.txt
discord.py==1.7.3
spotipy==2.18.0
ffmpeg-python==0.2.0
